+++ src/core_modules/capture-core/components/Pages/MainPage/EventsList/ListWrapper/EventsListWrapper.component.js
@@ -12,7 +12,7 @@ import ColumnSelector from './ColumnSelector.component';
 import OptionSet from '../../../../../metaData/OptionSet/OptionSet';
 import withCustomEndCell from '../withCustomEndCell';
 import eventContentMenuSettings from '../EventContentMenu/eventContentMenuSettings';
-import DialogLoadingMask from '../../../../LoadingMasks/DialogLoadingMask.component';
+import { DialogLoadingMask } from '../../../../LoadingMasks';
 
 import List from '../../../../List/OnlineList/List.component';
 import ListWrapperMenu from './ListWrapperMenu.component';
